v6.5.1

+++ bus-gitlab/src/main/java/org/aoju/bus/gitlab/ApplicationSettingsApi.java
@@ -61,7 +61,7 @@ public class ApplicationSettingsApi extends AbstractApi {
             String fieldName = fieldNames.next();
             switch (fieldName) {
                 case "id":
-                    appSettings.setId(root.path(fieldName).asInt());
+                    appSettings.setId(root.path(fieldName).asLong());
                     break;
 
                 case "created_at":
