+++ bus-gitlab/src/main/java/org/aoju/bus/gitlab/hooks/web/AbstractPushEvent.java
@@ -37,12 +37,12 @@ public abstract class AbstractPushEvent {
     private String before;
     private String ref;
     private String checkoutSha;
-    private Integer userId;
+    private Long userId;
     private String userName;
     private String userUsername;
     private String userEmail;
     private String userAvatar;
-    private Integer projectId;
+    private Long projectId;
     private EventProject project;
     private EventRepository repository;
     private List<EventCommit> commits;
@@ -91,11 +91,11 @@ public abstract class AbstractPushEvent {
         this.checkoutSha = checkoutSha;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
@@ -131,11 +131,11 @@ public abstract class AbstractPushEvent {
         this.userAvatar = userAvatar;
     }
 
-    public Integer getProjectId() {
+    public Long getProjectId() {
         return projectId;
     }
 
-    public void setProjectId(Integer projectId) {
+    public void setProjectId(Long projectId) {
         this.projectId = projectId;
     }
