+++ bus-gitlab/src/main/java/org/aoju/bus/gitlab/models/AllowedTo.java
@@ -34,10 +34,10 @@ import org.aoju.bus.gitlab.GitLabApiForm;
 public class AllowedTo {
 
     private AccessLevel accessLevel;
-    private Integer userId;
-    private Integer groupId;
+    private Long userId;
+    private Long groupId;
 
-    public AllowedTo(AccessLevel accessLevel, Integer userId, Integer groupId) {
+    public AllowedTo(AccessLevel accessLevel, Long userId, Long groupId) {
         this.accessLevel = accessLevel;
         this.userId = userId;
         this.groupId = groupId;
@@ -48,12 +48,12 @@ public class AllowedTo {
         return (this);
     }
 
-    public AllowedTo withUserId(Integer userId) {
+    public AllowedTo withUserId(Long userId) {
         this.userId = userId;
         return (this);
     }
 
-    public AllowedTo withGroupId(Integer groupId) {
+    public AllowedTo withGroupId(Long groupId) {
         this.groupId = groupId;
         return (this);
     }
