+++ bus-gitlab/src/main/java/org/aoju/bus/gitlab/models/Runner.java
@@ -32,7 +32,7 @@ import org.aoju.bus.gitlab.support.JacksonJsonEnumHelper;
 
 public class Runner {
 
-    private Integer id;
+    private Long id;
     private String description;
     private Boolean active;
     private Boolean isShared;
@@ -41,11 +41,11 @@ public class Runner {
     private RunnerStatus status;
     private String ipAddress;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -105,7 +105,7 @@ public class Runner {
         this.ipAddress = ipAddress;
     }
 
-    public Runner withId(Integer id) {
+    public Runner withId(Long id) {
         this.id = id;
         return this;
     }
