+++ bus-gitlab/src/main/java/org/aoju/bus/gitlab/models/SshKey.java
@@ -32,18 +32,18 @@ import java.util.Date;
 
 public class SshKey {
 
-    private Integer id;
+    private Long id;
     private String title;
     private String key;
     private Date createdAt;
 
-    private Integer userId;
+    private Long userId;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -72,11 +72,11 @@ public class SshKey {
     }
 
     @JsonIgnore
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
